requirements.txt
pytesseract
PyPDF2
python-docx
fastapi
uvicorn
httpx
pytest-asyncio
python-multipart
reportlab